+++ CMakeLists.txt
@@ -29,5 +29,4 @@ INCLUDE_DIRECTORIES(${ZEROMQ_INCLUDE_DIRS})
 LINK_DIRECTORIES(${BOOST_LINK_DIRS})
 LINK_DIRECTORIES(${ZEROMQ_LINK_DIRS})
 
-ADD_SUBDIRECTORY(kv_store)
-
+ADD_SUBDIRECTORY(src)
