+++ k8s/Dockerfile.kops
@@ -1,6 +1,6 @@
 FROM ubuntu:14.04
 
-MAINTAINER Chenggang Wu <cgwu0530@gmail.com> version: 0.1
+MAINTAINER Vikram Sreekanti <vsreekanti@gmail..com> version: 0.1
 
 USER root
 
@@ -33,9 +33,6 @@ RUN mv ./kubectl /usr/local/bin/kubectl
 RUN apt-get install -y git
 
 RUN git clone https://github.com/cw75/tiered-storage /tiered-storage
-# TODO: eventually remove
-RUN rm -rf /tiered-storage/k8s
-COPY ./ /tiered-storage/k8s
 
 # install jq
 RUN apt-get install -y jq
@@ -43,6 +40,6 @@ RUN apt-get install -y jq
 # make kube root dir
 RUN mkdir /root/.kube
 
-COPY ./start_kops.sh /
+RUN cp tiered-storage/k8s/start_kops.sh .
 
 CMD sh start_kops.sh
