+++ src/CMakeLists.txt
@@ -2,4 +2,4 @@ cmake_minimum_required (VERSION 2.6)
 
 INCLUDE_DIRECTORIES(include)
 
-ADD_SUBDIRECTORY(lww_kvs)
+ADD_SUBDIRECTORY(bedrock)
