+++ src/bedrock/kvs_benchmark.cpp
