+++ src/bedrock/kvs_monitoring.cpp
