+++ src/bedrock/kvs_proxy.cpp
