+++ src/include/utils/server_utility.h
@@ -5,6 +5,7 @@
 #include "message.pb.h"
 #include "../zmq/socket_cache.h"
 #include "../zmq/zmq_util.h"
+#include <yaml-cpp/yaml.h>
 
 using namespace std;
 
@@ -16,9 +17,6 @@ using namespace std;
 
 // Define the gossip period (frequency)
 #define PERIOD 10000000
-
-// Define the locatioon of the conf file with the ebs root path
-#define EBS_ROOT_FILE "conf/ebs_root.txt"
 
 // TODO: reconsider type names here
 typedef KV_Store<string, RC_KVS_PairLattice<string>> Database;
@@ -35,15 +33,19 @@ public:
 
 class Memory_Serializer : public Serializer {
   Database* kvs_;
+
 public:
   Memory_Serializer(Database* kvs): kvs_(kvs) {}
+
   RC_KVS_PairLattice<string> get(const string& key, unsigned& err_number) {
     return kvs_->get(key, err_number);
   }
+
   bool put(const string& key, const string& value, const unsigned& timestamp) {
     timestamp_value_pair<string> p = timestamp_value_pair<string>(timestamp, value);
     return kvs_->put(key, RC_KVS_PairLattice<string>(p));
   }
+
   void remove(const string& key) {
     kvs_->remove(key);
   }
@@ -52,24 +54,24 @@ public:
 class EBS_Serializer : public Serializer {
   unsigned tid_;
   string ebs_root_;
+
 public:
   EBS_Serializer(unsigned& tid): tid_(tid) {
-    ifstream address;
+    YAML::Node conf = YAML::LoadFile("conf/config.yml");
 
-    address.open(EBS_ROOT_FILE);
-    std::getline(address, ebs_root_);
-    address.close();
+    string monitoring_address = conf["ebs"].as<string>();
 
     if (ebs_root_.back() != '/') {
       ebs_root_ += "/";
     }
   }
+
   RC_KVS_PairLattice<string> get(const string& key, unsigned& err_number) {
     RC_KVS_PairLattice<string> res;
-
     communication::Payload pl;
-    string fname = ebs_root_ + "ebs_" + to_string(tid_) + "/" + key;
+
     // open a new filestream for reading in a binary
+    string fname = ebs_root_ + "ebs_" + to_string(tid_) + "/" + key;
     fstream input(fname, ios::in | ios::binary);
 
     if (!input) {
@@ -82,6 +84,7 @@ public:
     }
     return res;
   }
+
   bool put(const string& key, const string& value, const unsigned& timestamp) {
     bool replaced = false;
     timestamp_value_pair<string> p = timestamp_value_pair<string>(timestamp, value);
@@ -108,21 +111,27 @@ public:
       // get the existing value that we have and merge
       RC_KVS_PairLattice<string> l = RC_KVS_PairLattice<string>(timestamp_value_pair<string>(pl_orig.timestamp(), pl_orig.value()));
       replaced = l.Merge(p);
+
       if (replaced) {
         // set the payload's data to the merged values of the value and timestamp
         pl.set_timestamp(l.reveal().timestamp);
         pl.set_value(l.reveal().value);
+
         // write out the new payload.
         fstream output(fname, ios::out | ios::trunc | ios::binary);
+
         if (!pl.SerializeToOstream(&output)) {
           cerr << "Failed to write payload\n";
         }
       }
     }
+
     return replaced;
   }
+
   void remove(const string& key) {
     string fname = ebs_root_ + "ebs_" + to_string(tid_) + "/" + key;
+
     if(std::remove(fname.c_str()) != 0) {
       cout << "Error deleting file";
     }
@@ -157,10 +166,10 @@ struct pending_gossip {
 
 // form the timestamp given a time and a thread id
 unsigned long long generate_timestamp(unsigned long long time, unsigned tid) {
-    unsigned pow = 10;
-    while(tid >= pow)
-        pow *= 10;
-    return time * pow + tid;
+  unsigned pow = 10;
+  while(tid >= pow)
+    pow *= 10;
+  return time * pow + tid;
 }
 
 #endif
